refactor: move BaseConnetion in its own module
This commit is brought you by: flight delayed 6 hours for technical failure.
