+++ psycopg/psycopg/__init__.py
@@ -19,12 +19,13 @@ from .errors import InternalError, ProgrammingError, NotSupportedError
 from ._column import Column
 from .conninfo import ConnectionInfo
 from ._pipeline import Pipeline, AsyncPipeline
-from .connection import BaseConnection, Connection, Notify
+from .connection import Connection
 from .transaction import Rollback, Transaction, AsyncTransaction
 from .cursor_async import AsyncCursor
 from .server_cursor import AsyncServerCursor, ServerCursor
 from .client_cursor import AsyncClientCursor, ClientCursor
 from .raw_cursor import AsyncRawCursor, RawCursor
+from ._connection_base import BaseConnection, Notify
 from .connection_async import AsyncConnection
 
 from . import dbapi20
